(refactor) core: decompose CampaignRepository into focused repositories
Extract exclude-list and statistics methods into dedicated repository
classes to improve cohesion and reduce CampaignRepository from 1059
to 582 lines.

- CampaignExcludeListRepository: getExcludeList, addToExcludeList,
  removeFromExcludeList
- CampaignStatisticsRepository: getStatistics, resetForRerun
- CampaignRepository retains: CRUD, creation support, action chain ops
- Update all MCP, CLI, and service consumers and their tests
