+++ packages/core/src/db/repositories/campaign.test.ts
@@ -8,7 +8,6 @@ import { DatabaseClient } from "../client.js";
 import {
   ActionNotFoundError,
   CampaignNotFoundError,
-  ExcludeListNotFoundError,
   NoNextActionError,
 } from "../errors.js";
 import { openFixture } from "../testing/open-fixture.js";
@@ -422,140 +421,8 @@ describe("CampaignRepository", () => {
     });
   });
 
-  describe("resetForRerun", () => {
-    it("resets person state for re-run", () => {
-      // Initial state: person 1 has state=2 (processed), person 3 has state=1 (queued)
-      const initialTargets = db
-        .prepare(
-          "SELECT person_id, state FROM action_target_people WHERE action_id = 1",
-        )
-        .all() as Array<{ person_id: number; state: number }>;
-
-      const initialPerson1 = initialTargets.find((t) => t.person_id === 1);
-      expect(initialPerson1).toBeDefined();
-      expect(initialPerson1?.state).toBe(2);
-
-      // Reset person 1
-      repo.resetForRerun(1, [1]);
-
-      // Verify state reset to 1 (queued)
-      const afterTargets = db
-        .prepare(
-          "SELECT person_id, state FROM action_target_people WHERE action_id = 1",
-        )
-        .all() as Array<{ person_id: number; state: number }>;
-
-      const afterPerson1 = afterTargets.find((t) => t.person_id === 1);
-      expect(afterPerson1).toBeDefined();
-      expect(afterPerson1?.state).toBe(1);
-    });
-
-    it("resets campaign history for re-run", () => {
-      // Initial state: person 1 has result_status=1 (success)
-      const initialHistory = db
-        .prepare(
-          "SELECT result_status FROM person_in_campaigns_history WHERE campaign_id = 1 AND person_id = 1",
-        )
-        .get() as { result_status: number } | undefined;
-
-      expect(initialHistory).toBeDefined();
-      expect(initialHistory?.result_status).toBe(1);
-
-      // Reset person 1
-      repo.resetForRerun(1, [1]);
-
-      // Verify result_status reset to -999
-      const afterHistory = db
-        .prepare(
-          "SELECT result_status FROM person_in_campaigns_history WHERE campaign_id = 1 AND person_id = 1",
-        )
-        .get() as { result_status: number } | undefined;
-
-      expect(afterHistory).toBeDefined();
-      expect(afterHistory?.result_status).toBe(-999);
-    });
-
-    it("deletes old action results and related records", () => {
-      // Verify initial state has results
-      const initialResults = db
-        .prepare(
-          "SELECT id FROM action_results WHERE action_version_id = 1 AND person_id = 1",
-        )
-        .all() as Array<{ id: number }>;
-      expect(initialResults).toHaveLength(1);
-
-      const initialFlags = db
-        .prepare(
-          "SELECT id FROM action_result_flags WHERE action_result_id = 1",
-        )
-        .all();
-      expect(initialFlags).toHaveLength(1);
-
-      const initialMessages = db
-        .prepare(
-          "SELECT id FROM action_result_messages WHERE action_result_id = 1",
-        )
-        .all();
-      expect(initialMessages).toHaveLength(1);
-
-      // Reset person 1
-      repo.resetForRerun(1, [1]);
-
-      // Verify all deleted
-      const afterResults = db
-        .prepare(
-          "SELECT id FROM action_results WHERE action_version_id = 1 AND person_id = 1",
-        )
-        .all();
-      expect(afterResults).toHaveLength(0);
-
-      const afterFlags = db
-        .prepare(
-          "SELECT id FROM action_result_flags WHERE action_result_id = 1",
-        )
-        .all();
-      expect(afterFlags).toHaveLength(0);
-
-      const afterMessages = db
-        .prepare(
-          "SELECT id FROM action_result_messages WHERE action_result_id = 1",
-        )
-        .all();
-      expect(afterMessages).toHaveLength(0);
-    });
-
-    it("handles multiple persons in a single reset", () => {
-      repo.resetForRerun(1, [1, 3]);
-
-      // Both persons should have state=1
-      const targets = db
-        .prepare(
-          "SELECT person_id, state FROM action_target_people WHERE action_id = 1",
-        )
-        .all() as Array<{ person_id: number; state: number }>;
-
-      const stateByPerson = new Map(targets.map((t) => [t.person_id, t.state]));
-      expect(stateByPerson.get(1)).toBe(1);
-      expect(stateByPerson.get(3)).toBe(1);
-    });
-
-    it("handles empty person list gracefully", () => {
-      // Should not throw
-      expect(() => repo.resetForRerun(1, [])).not.toThrow();
-    });
-
-    it("throws CampaignNotFoundError for missing campaign", () => {
-      expect(() => repo.resetForRerun(999, [1])).toThrow(CampaignNotFoundError);
-    });
-
-    it("handles campaign with no actions gracefully", () => {
-      // Campaign 3 has no actions
-      expect(() => repo.resetForRerun(3, [1])).not.toThrow();
-    });
-  });
-
   describe("moveToNextAction", () => {
-    // Campaign 5 has 3 actions: action 5 (VisitAndExtract) → action 6 (Waiter) → action 7 (InvitePerson)
+    // Campaign 5 has 3 actions: action 5 (VisitAndExtract) -> action 6 (Waiter) -> action 7 (InvitePerson)
     // Person 1 is queued (state=1) in action 5, person 3 is processed (state=2) in action 5
 
     it("moves person from current action to next action", () => {
@@ -674,383 +541,6 @@ describe("CampaignRepository", () => {
     it("handles empty person list gracefully", () => {
       const result = repo.moveToNextAction(5, 5, []);
       expect(result.nextActionId).toBe(0);
-    });
-  });
-
-  describe("getExcludeList", () => {
-    // Campaign 1 has campaign-level exclude list with person 2 pre-populated
-    // Campaign 5 has campaign-level exclude list (empty) and action-level lists (empty)
-
-    it("returns campaign-level exclude list", () => {
-      const entries = repo.getExcludeList(1);
-
-      expect(entries).toEqual([{ personId: 2 }]);
-    });
-
-    it("returns empty list when no people excluded", () => {
-      const entries = repo.getExcludeList(5);
-
-      expect(entries).toEqual([]);
-    });
-
-    it("returns action-level exclude list", () => {
-      const entries = repo.getExcludeList(1, 1);
-
-      // Action 1's exclude list (collection 2) has no people
-      expect(entries).toEqual([]);
-    });
-
-    it("throws CampaignNotFoundError for missing campaign", () => {
-      expect(() => repo.getExcludeList(999)).toThrow(CampaignNotFoundError);
-    });
-
-    it("throws ActionNotFoundError for action not in campaign", () => {
-      expect(() => repo.getExcludeList(1, 999)).toThrow(ActionNotFoundError);
-    });
-
-    it("throws ExcludeListNotFoundError when campaign has no exclude list chain", () => {
-      // Campaign 2 has no campaign_versions entry
-      expect(() => repo.getExcludeList(2)).toThrow(ExcludeListNotFoundError);
-    });
-
-    it("returns entries ordered by person_id", () => {
-      // Add multiple people to campaign 1's exclude list
-      db.exec(`
-        INSERT INTO collection_people (collection_id, person_id) VALUES (1, 3);
-        INSERT INTO collection_people (collection_id, person_id) VALUES (1, 1);
-      `);
-
-      const entries = repo.getExcludeList(1);
-
-      expect(entries.map((e) => e.personId)).toEqual([1, 2, 3]);
-    });
-  });
-
-  describe("addToExcludeList", () => {
-    it("adds people to campaign-level exclude list", () => {
-      const added = repo.addToExcludeList(1, [1, 3]);
-
-      expect(added).toBe(2);
-
-      // Verify in database
-      const entries = repo.getExcludeList(1);
-      expect(entries.map((e) => e.personId)).toContain(1);
-      expect(entries.map((e) => e.personId)).toContain(3);
-    });
-
-    it("skips already-excluded people", () => {
-      // Person 2 is already in campaign 1's exclude list
-      const added = repo.addToExcludeList(1, [2, 3]);
-
-      expect(added).toBe(1); // Only person 3 was newly added
-
-      const entries = repo.getExcludeList(1);
-      expect(entries).toHaveLength(2); // Person 2 (existing) + person 3 (new)
-    });
-
-    it("adds people to action-level exclude list", () => {
-      const added = repo.addToExcludeList(1, [1], 1);
-
-      expect(added).toBe(1);
-
-      const entries = repo.getExcludeList(1, 1);
-      expect(entries).toEqual([{ personId: 1 }]);
-    });
-
-    it("returns 0 for empty person list", () => {
-      const added = repo.addToExcludeList(1, []);
-
-      expect(added).toBe(0);
-    });
-
-    it("throws CampaignNotFoundError for missing campaign", () => {
-      expect(() => repo.addToExcludeList(999, [1])).toThrow(
-        CampaignNotFoundError,
-      );
-    });
-
-    it("throws ActionNotFoundError for action not in campaign", () => {
-      expect(() => repo.addToExcludeList(1, [1], 999)).toThrow(
-        ActionNotFoundError,
-      );
-    });
-
-    it("throws ExcludeListNotFoundError when exclude list chain missing", () => {
-      expect(() => repo.addToExcludeList(2, [1])).toThrow(
-        ExcludeListNotFoundError,
-      );
-    });
-  });
-
-  describe("removeFromExcludeList", () => {
-    it("removes people from campaign-level exclude list", () => {
-      // Person 2 is in campaign 1's exclude list
-      const removed = repo.removeFromExcludeList(1, [2]);
-
-      expect(removed).toBe(1);
-
-      const entries = repo.getExcludeList(1);
-      expect(entries).toEqual([]);
-    });
-
-    it("returns 0 for people not in the list", () => {
-      const removed = repo.removeFromExcludeList(1, [999]);
-
-      expect(removed).toBe(0);
-    });
-
-    it("removes people from action-level exclude list", () => {
-      // First add a person to action 1's exclude list
-      repo.addToExcludeList(1, [1], 1);
-
-      const removed = repo.removeFromExcludeList(1, [1], 1);
-
-      expect(removed).toBe(1);
-
-      const entries = repo.getExcludeList(1, 1);
-      expect(entries).toEqual([]);
-    });
-
-    it("handles mixed present and absent people", () => {
-      // Person 2 is in the list, person 999 is not
-      const removed = repo.removeFromExcludeList(1, [2, 999]);
-
-      expect(removed).toBe(1);
-    });
-
-    it("returns 0 for empty person list", () => {
-      const removed = repo.removeFromExcludeList(1, []);
-
-      expect(removed).toBe(0);
-    });
-
-    it("throws CampaignNotFoundError for missing campaign", () => {
-      expect(() => repo.removeFromExcludeList(999, [1])).toThrow(
-        CampaignNotFoundError,
-      );
-    });
-
-    it("throws ActionNotFoundError for action not in campaign", () => {
-      expect(() => repo.removeFromExcludeList(1, [1], 999)).toThrow(
-        ActionNotFoundError,
-      );
-    });
-
-    it("throws ExcludeListNotFoundError when exclude list chain missing", () => {
-      expect(() => repo.removeFromExcludeList(2, [1])).toThrow(
-        ExcludeListNotFoundError,
-      );
-    });
-  });
-
-  describe("getStatistics", () => {
-    // Campaign 1 fixture data:
-    //   Action 1 ("Send Welcome Message", MessageToPerson) with 4 results:
-    //     person 1: result=1 (successful)
-    //     person 3: result=2 (replied)
-    //     person 2: result=-1 (failed)
-    //     person 4: result=-2 (skipped)
-    //   Error flags (code != NULL):
-    //     code=100, is_exception=0, who_to_blame='LinkedIn' (x2)
-    //     code=200, is_exception=1, who_to_blame='LH' (x1)
-
-    it("returns per-action breakdown for a campaign", () => {
-      const stats = repo.getStatistics(1);
-
-      expect(stats.campaignId).toBe(1);
-      expect(stats.actions).toHaveLength(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      expect(action).toMatchObject({
-        actionId: 1,
-        actionName: "Send Welcome Message",
-        actionType: "MessageToPerson",
-      });
-    });
-
-    it("counts successful/replied/failed/skipped correctly", () => {
-      const stats = repo.getStatistics(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      expect(action?.successful).toBe(1);
-      expect(action?.replied).toBe(1);
-      expect(action?.failed).toBe(1);
-      expect(action?.skipped).toBe(1);
-      expect(action?.total).toBe(4);
-    });
-
-    it("calculates success rate correctly", () => {
-      const stats = repo.getStatistics(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      // successRate = round(((successful + replied) / total) * 1000) / 10
-      // = round(((1 + 1) / 4) * 1000) / 10 = round(500) / 10 = 50
-      expect(action?.successRate).toBe(50);
-    });
-
-    it("returns top errors with blame attribution", () => {
-      const stats = repo.getStatistics(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      const errors = action?.topErrors ?? [];
-
-      expect(errors).toHaveLength(2);
-      expect(errors.at(0)).toMatchObject({
-        code: 100,
-        count: 2,
-        isException: false,
-        whoToBlame: "LinkedIn",
-      });
-      expect(errors.at(1)).toMatchObject({
-        code: 200,
-        count: 1,
-        isException: true,
-        whoToBlame: "LH",
-      });
-    });
-
-    it("orders top errors by count descending", () => {
-      const stats = repo.getStatistics(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      const errors = action?.topErrors ?? [];
-
-      for (let i = 1; i < errors.length; i++) {
-        expect(errors[i - 1]?.count).toBeGreaterThanOrEqual(
-          errors[i]?.count ?? 0,
-        );
-      }
-    });
-
-    it("respects maxErrors option", () => {
-      const stats = repo.getStatistics(1, { maxErrors: 1 });
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      const errors = action?.topErrors ?? [];
-
-      expect(errors).toHaveLength(1);
-      expect(errors.at(0)?.code).toBe(100);
-    });
-
-    it("filters by actionId when provided", () => {
-      const stats = repo.getStatistics(1, { actionId: 1 });
-
-      expect(stats.actions).toHaveLength(1);
-      expect(stats.actions.at(0)?.actionId).toBe(1);
-    });
-
-    it("throws CampaignNotFoundError for missing campaign", () => {
-      expect(() => repo.getStatistics(999)).toThrow(CampaignNotFoundError);
-    });
-
-    it("throws ActionNotFoundError for invalid actionId", () => {
-      expect(() => repo.getStatistics(1, { actionId: 999 })).toThrow(
-        ActionNotFoundError,
-      );
-    });
-
-    it("handles campaign with no results", () => {
-      // Campaign 3 has no actions (archived)
-      const stats = repo.getStatistics(3);
-
-      expect(stats.campaignId).toBe(3);
-      expect(stats.actions).toHaveLength(0);
-      expect(stats.totals).toMatchObject({
-        successful: 0,
-        replied: 0,
-        failed: 0,
-        skipped: 0,
-        total: 0,
-        successRate: 0,
-      });
-    });
-
-    it("aggregates totals across all actions", () => {
-      const stats = repo.getStatistics(1);
-
-      expect(stats.totals).toMatchObject({
-        successful: 1,
-        replied: 1,
-        failed: 1,
-        skipped: 1,
-        total: 4,
-        successRate: 50,
-      });
-    });
-
-    it("includes first and last result timestamps", () => {
-      const stats = repo.getStatistics(1);
-
-      const action = stats.actions.at(0);
-      expect(action).toBeDefined();
-      expect(action?.firstResultAt).toBe("2025-01-15T12:30:00.000Z");
-      expect(action?.lastResultAt).toBe("2025-01-15T14:00:00.000Z");
-    });
-  });
-
-  describe("resetForRerun — rollback", () => {
-    it("rolls back transaction on error and preserves database state", () => {
-      // Snapshot initial state
-      const initialTargets = db
-        .prepare(
-          "SELECT person_id, state FROM action_target_people WHERE action_id = 1 ORDER BY person_id",
-        )
-        .all() as Array<{ person_id: number; state: number }>;
-
-      const initialHistory = db
-        .prepare(
-          "SELECT person_id, result_status FROM person_in_campaigns_history WHERE campaign_id = 1 ORDER BY person_id",
-        )
-        .all() as Array<{ person_id: number; result_status: number }>;
-
-      const initialResults = db
-        .prepare(
-          "SELECT id FROM action_results WHERE action_version_id = 1 ORDER BY id",
-        )
-        .all() as Array<{ id: number }>;
-
-      // Sabotage a write statement to force an error mid-transaction
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any -- private method access for rollback test
-      const stmts = (repo as any).getWriteStatements();
-      const originalRun = stmts.resetHistory.run.bind(stmts.resetHistory);
-      stmts.resetHistory.run = () => {
-        throw new Error("Simulated DB error");
-      };
-
-      // Attempt reset — should throw the simulated error
-      expect(() => repo.resetForRerun(1, [1])).toThrow("Simulated DB error");
-
-      // Restore original to avoid side effects
-      stmts.resetHistory.run = originalRun;
-
-      // Verify database state is unchanged (ROLLBACK worked)
-      const afterTargets = db
-        .prepare(
-          "SELECT person_id, state FROM action_target_people WHERE action_id = 1 ORDER BY person_id",
-        )
-        .all() as Array<{ person_id: number; state: number }>;
-
-      const afterHistory = db
-        .prepare(
-          "SELECT person_id, result_status FROM person_in_campaigns_history WHERE campaign_id = 1 ORDER BY person_id",
-        )
-        .all() as Array<{ person_id: number; result_status: number }>;
-
-      const afterResults = db
-        .prepare(
-          "SELECT id FROM action_results WHERE action_version_id = 1 ORDER BY id",
-        )
-        .all() as Array<{ id: number }>;
-
-      expect(afterTargets).toEqual(initialTargets);
-      expect(afterHistory).toEqual(initialHistory);
-      expect(afterResults).toEqual(initialResults);
     });
   });
 });
