+++ packages/core/src/db/repositories/index.ts
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 // Copyright (C) 2025 Alexey Pelykh
 
+export { CampaignExcludeListRepository } from "./campaign-exclude-list.js";
 export { CampaignRepository } from "./campaign.js";
+export { CampaignStatisticsRepository } from "./campaign-statistics.js";
 export { MessageRepository } from "./message.js";
 export { ProfileRepository } from "./profile.js";
