+++ packages/mcp/src/tools/campaign-exclude-list.test.ts
@@ -9,14 +9,14 @@ vi.mock("@lhremote/core", async (importOriginal) => {
     ...actual,
     resolveAccount: vi.fn(),
     withDatabase: vi.fn(),
-    CampaignRepository: vi.fn(),
+    CampaignExcludeListRepository: vi.fn(),
   };
 });
 
 import {
   ActionNotFoundError,
   CampaignNotFoundError,
-  CampaignRepository,
+  CampaignExcludeListRepository,
   type DatabaseContext,
   ExcludeListNotFoundError,
   resolveAccount,
@@ -31,10 +31,10 @@ function mockCampaignRepo() {
   const getExcludeList = vi
     .fn()
     .mockReturnValue([{ personId: 1 }, { personId: 2 }]);
-  vi.mocked(CampaignRepository).mockImplementation(function () {
+  vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
     return {
       getExcludeList,
-    } as unknown as CampaignRepository;
+    } as unknown as CampaignExcludeListRepository;
   });
   return { getExcludeList };
 }
@@ -182,12 +182,12 @@ describe("registerCampaignExcludeList", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         getExcludeList: vi.fn().mockImplementation(() => {
           throw new CampaignNotFoundError(999);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-list");
@@ -215,12 +215,12 @@ describe("registerCampaignExcludeList", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         getExcludeList: vi.fn().mockImplementation(() => {
           throw new ActionNotFoundError(5, 10);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-list");
@@ -249,12 +249,12 @@ describe("registerCampaignExcludeList", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         getExcludeList: vi.fn().mockImplementation(() => {
           throw new ExcludeListNotFoundError("campaign", 10);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-list");
