+++ packages/mcp/src/tools/campaign-exclude-remove.test.ts
@@ -9,14 +9,14 @@ vi.mock("@lhremote/core", async (importOriginal) => {
     ...actual,
     resolveAccount: vi.fn(),
     withDatabase: vi.fn(),
-    CampaignRepository: vi.fn(),
+    CampaignExcludeListRepository: vi.fn(),
   };
 });
 
 import {
   ActionNotFoundError,
   CampaignNotFoundError,
-  CampaignRepository,
+  CampaignExcludeListRepository,
   type DatabaseContext,
   ExcludeListNotFoundError,
   resolveAccount,
@@ -29,10 +29,10 @@ import { createMockServer } from "./testing/mock-server.js";
 
 function mockCampaignRepo() {
   const removeFromExcludeList = vi.fn().mockReturnValue(1);
-  vi.mocked(CampaignRepository).mockImplementation(function () {
+  vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
     return {
       removeFromExcludeList,
-    } as unknown as CampaignRepository;
+    } as unknown as CampaignExcludeListRepository;
   });
   return { removeFromExcludeList };
 }
@@ -154,12 +154,12 @@ describe("registerCampaignExcludeRemove", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         removeFromExcludeList: vi.fn().mockImplementation(() => {
           throw new CampaignNotFoundError(999);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-remove");
@@ -188,12 +188,12 @@ describe("registerCampaignExcludeRemove", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         removeFromExcludeList: vi.fn().mockImplementation(() => {
           throw new ActionNotFoundError(5, 10);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-remove");
@@ -223,12 +223,12 @@ describe("registerCampaignExcludeRemove", () => {
     vi.mocked(withDatabase).mockImplementation(async (_accountId, callback) =>
       callback({ accountId: 1, db: {} } as unknown as DatabaseContext),
     );
-    vi.mocked(CampaignRepository).mockImplementation(function () {
+    vi.mocked(CampaignExcludeListRepository).mockImplementation(function () {
       return {
         removeFromExcludeList: vi.fn().mockImplementation(() => {
           throw new ExcludeListNotFoundError("campaign", 10);
         }),
-      } as unknown as CampaignRepository;
+      } as unknown as CampaignExcludeListRepository;
     });
 
     const handler = getHandler("campaign-exclude-remove");
